requirements.txt
# generated by rye
# use `rye lock` or `rye sync` to update this lockfile
#
# last locked with the following flags:
#   pre: false
#   features: []
#   all-features: false
#   with-sources: false

absl-py==2.1.0
    # via tensorboard
aiohttp==3.9.3
    # via fsspec
aiosignal==1.3.1
    # via aiohttp
annotated-types==0.6.0
    # via pydantic
anyio==4.3.0
    # via starlette
async-timeout==4.0.3
    # via aiohttp
attrs==23.2.0
    # via aiohttp
blosc2==2.5.1
    # via tables
cachetools==5.3.3
    # via google-auth
certifi==2024.2.2
    # via requests
charset-normalizer==3.3.2
    # via requests
colorama==0.4.6
    # via tqdm
exceptiongroup==1.2.0
    # via anyio
fastapi==0.110.0
    # via smaat-unet
filelock==3.13.3
    # via torch
frozenlist==1.4.1
    # via aiohttp
    # via aiosignal
fsspec==2024.3.1
    # via lightning
    # via pytorch-lightning
    # via torch
google-auth==2.29.0
    # via google-auth-oauthlib
    # via tensorboard
google-auth-oauthlib==1.0.0
    # via tensorboard
grpcio==1.62.1
    # via tensorboard
h5py==3.11.0
    # via smaat-unet
idna==3.6
    # via anyio
    # via requests
    # via yarl
jinja2==3.1.3
    # via torch
lightning==2.2.1
    # via smaat-unet
lightning-utilities==0.11.2
    # via lightning
    # via pytorch-lightning
    # via torchmetrics
markdown==3.6
    # via tensorboard
markupsafe==2.1.5
    # via jinja2
    # via werkzeug
mpmath==1.3.0
    # via sympy
msgpack==1.0.8
    # via blosc2
multidict==6.0.5
    # via aiohttp
    # via yarl
ndindex==1.8
    # via blosc2
networkx==3.2.1
    # via torch
numexpr==2.9.0
    # via tables
numpy==1.26.4
    # via blosc2
    # via h5py
    # via lightning
    # via numexpr
    # via pandas
    # via pytorch-lightning
    # via tables
    # via tensorboard
    # via torchmetrics
    # via torchvision
oauthlib==3.2.2
    # via requests-oauthlib
packaging==24.0
    # via lightning
    # via lightning-utilities
    # via pytorch-lightning
    # via tables
    # via torchmetrics
pandas==1.5.3
    # via smaat-unet
pillow==10.2.0
    # via torchvision
protobuf==5.26.1
    # via tensorboard
py-cpuinfo==9.0.0
    # via blosc2
    # via tables
pyasn1==0.6.0
    # via pyasn1-modules
    # via rsa
pyasn1-modules==0.4.0
    # via google-auth
pydantic==2.6.4
    # via fastapi
pydantic-core==2.16.3
    # via pydantic
python-dateutil==2.9.0.post0
    # via pandas
pytorch-lightning==2.2.1
    # via lightning
pytz==2024.1
    # via pandas
pyyaml==6.0.1
    # via lightning
    # via pytorch-lightning
requests==2.31.0
    # via requests-oauthlib
    # via tensorboard
requests-oauthlib==2.0.0
    # via google-auth-oauthlib
rsa==4.9
    # via google-auth
six==1.16.0
    # via python-dateutil
sniffio==1.3.1
    # via anyio
starlette==0.36.3
    # via fastapi
sympy==1.12
    # via torch
tables==3.9.2
    # via smaat-unet
tensorboard==2.13.0
    # via smaat-unet
tensorboard-data-server==0.7.2
    # via tensorboard
torch==2.3.0+cu121
    # via lightning
    # via pytorch-lightning
    # via smaat-unet
    # via torchmetrics
    # via torchvision
torchmetrics==1.3.2
    # via lightning
    # via pytorch-lightning
torchsummary==1.5.1
    # via smaat-unet
torchvision==0.18.0+cu121
    # via smaat-unet
tqdm==4.65.2
    # via lightning
    # via pytorch-lightning
    # via smaat-unet
    # via anyio
    # via fastapi
    # via lightning
    # via lightning-utilities
    # via pydantic
    # via pydantic-core
    # via pytorch-lightning
    # via torch
urllib3==2.2.1
    # via requests
werkzeug==3.0.1
    # via tensorboard
wheel==0.43.0
    # via tensorboard
yarl==1.9.4
    # via aiohttp
setuptools==69.2.0
    # via lightning-utilities
    # via tensorboard

matplotlib~=3.8.4
tifffile~=2024.5.3